refactor: replace ui with atoms
