+++ apps/client/next.config.js
@@ -1,4 +1,4 @@
-const withTM = require("next-transpile-modules")(["ui"]);
+const withTM = require("next-transpile-modules")(["atoms"]);
 
 module.exports = withTM({
   reactStrictMode: true,
