+++ apps/client/package.json
@@ -15,7 +15,7 @@
     "next": "13.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "ui": "workspace:*"
+    "atoms": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "^18.0.0",
