--- packages/ui/Button.tsx
@@ -1,4 +1,0 @@
-import * as React from "react";
-export const Button = () => {
-  return <button>Boop</button>;
-};
